docs(nuxtjs): add hightlight components and global Markdown
