+++ nuxtjs.org/components/Highlight.vue
@@ -1,0 +1,3 @@
+<template>
+  <span class="text-primary"><Markdown unwrap="p" /></span>
+</template>
