+++ nuxtjs.org/components/HomeFeatures.vue
@@ -29,10 +29,8 @@
 
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
-import { Markdown } from '~docus/utils'
 
 export default defineComponent({
-  components: { Markdown },
   props: {
     category: {
       type: String,
