+++ src/core/runtime/components/index.ts
@@ -1,0 +1,7 @@
+import Vue from 'vue'
+
+import Markdown from '~docus/components/Markdown'
+import DocusContent from '~docus/components/DocusContent'
+
+Vue.component('Markdown', Markdown)
+Vue.component('DocusContent', DocusContent)
