+++ src/defaultTheme/components/atoms/Alert.vue
@@ -10,10 +10,8 @@
 
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
-import { Markdown } from '~docus/utils'
 
 export default defineComponent({
-  components: { Markdown },
   props: {
     /**
      * @values info, success, warning, danger
