+++ src/defaultTheme/components/atoms/ButtonLink.vue
@@ -11,10 +11,8 @@
 
 <script>
 import { computed, defineComponent } from '@nuxtjs/composition-api'
-import { Markdown } from '~docus/utils'
 
 export default defineComponent({
-  components: { Markdown },
   props: {
     href: {
       type: String,
