+++ src/defaultTheme/components/organisms/BlockHero.vue
@@ -73,10 +73,8 @@
 
 <script>
 import { defineComponent } from '@nuxtjs/composition-api'
-import { Markdown } from '~docus/utils'
 
 export default defineComponent({
-  components: { Markdown },
   props: {
     cta: {
       type: Array,
